requirements.txt
Pillow
pyyaml
numpy